requirements.txt
streamlit==1.39.0
feedparser==6.0.11
pandas==2.2.3
altair==5.4.1
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
